+++ GoalMogul-iOS/src/Router.js
@@ -3,6 +3,7 @@ import {
   Scene,
   Router,
   Stack,
+  Tabs,
   Modal
 } from 'react-native-router-flux';
 
@@ -15,15 +16,19 @@ import IntroForm from './Registration/IntroForm';
 import AddProfilePic from './Registration/AddProfilePic';
 import Contacts from './Registration/Contacts';
 import ContactSync from './Registration/ContactSync';
-import CameraRollModal from './Registration/CameraRollModal';
-import Header from './Registration/Common/Header';
+
+/* Main App */
+import TabIcon from './Main/Common/TabIcon';
+import Home from './Main/Home/Home';
+
+import Profile from './Main/Profile/Profile';
 
 const RouterComponent = () => {
   return (
     <Router>
       <Modal>
         <Scene key="root" hideNavBar>
-          <Scene key="auth" hideNavBar>
+          <Scene key="auth" initial hideNavBar>
             <Scene key="login" component={Login} initial />
           </Scene>
 
@@ -51,6 +56,65 @@ const RouterComponent = () => {
               component={ContactSync}
             />
           </Stack>
+
+          {/* Main App */}
+
+          <Scene hideNavBar panHandlers={null}>
+            <Tabs
+              key="mainTabs"
+              hideNavBar
+              swipeEnabled
+              tabBarStyle={styles.tabBarStyle}
+              activeTintColor="#324a61"
+              inactiveTintColor="#b8c7cc"
+              tabs
+              showLabel={false}
+            >
+              <Stack
+                key="homeTab"
+                initial
+                icon={TabIcon}
+                hideNavBar
+              >
+                <Scene key="home" component={Home} initial hideNavBar />
+                <Scene key="profile" component={Profile} hideNavBar />
+              </Stack>
+
+              <Stack
+                key="goalTab"
+                icon={TabIcon}
+                hideNavBar
+              >
+                <Scene key="goal" component={Home} hideNavBar />
+              </Stack>
+
+              <Stack
+                key="notificationTab"
+                icon={TabIcon}
+                hideNavBar
+              >
+                <Scene key="notification" component={Home} />
+              </Stack>
+
+              <Stack
+                key="exploreTab"
+                icon={TabIcon}
+                hideNavBar
+              >
+                <Scene key="explore" component={Home} />
+              </Stack>
+
+              <Stack
+                key="chatTab"
+                icon={TabIcon}
+                hideNavBar
+              >
+                <Scene key="chat" component={Home} />
+              </Stack>
+
+            </Tabs>
+          </Scene>
+
         </Scene>
         {/*
           This model is deprecated. Using ImagePickerIOS instead.
@@ -68,6 +132,9 @@ const styles = {
   navBarStyle: {
     backgroundColor: '#34c0dd',
     borderBottomColor: 'transparent'
+  },
+  tabBarStyle: {
+
   }
 };