+++ GoalMogul-iOS/src/actions/types.js
@@ -5,9 +5,11 @@ export const LOGIN_USER_FAIL = 'login_user_fail';
 export const LOGIN_USER_LOADING = 'login_user_loading';
 
 /* Registration Constants */
+// General registration error 
 export const REGISTRATION_ERROR = 'registration_error';
 export const REGISTRATION_BACK = 'registration_back';
 export const REGISTRATION_ACCOUNT = 'registration_account';
+export const REGISTRATION_ACCOUNT_SUCCESS = 'registration_account_success';
 export const REGISTRATION_ACCOUNT_FORM_CHANGE = 'registration_account_form_change';
 
 export const REGISTRATION_LOGIN = 'registration_login';
@@ -17,6 +19,8 @@ export const REGISTRATION_INTRO_SKIP = 'registration_intro_skip';
 export const REGISTRATION_CONTACT = 'registration_contact';
 export const REGISTRATION_CONTACT_SKIP = 'registration_contact_skip';
 export const REGISTRATION_CONTACT_SYNC = 'registration_contact_sync';
+export const REGISTRATION_CONTACT_SYNC_SKIP = 'registration_contact_sync_skip';
+export const REGISTRATION_CONTACT_SYNC_DONE = 'registration_contact_sync_done';
 
 export const REGISTRATION_INTRO_FORM_CHANGE = 'registration_intro_form_change';